requirements.txt
# The following to be installed with apt-get or something:
# python-pip python-gdal python-xapian python-dev libpq-dev git-core postgresql-9.1-postgis

# These by pip
# General Python stuff
PIL >= 1.1.7
psycopg2
BeautifulSoup
matplotlib # ugly piecharts

# Django stuff
django-mptt==0.4.2
django-reversion==1.4
South
django-cms
sorl-thumbnail
django-tagging
django-modeltranslation

# Search
django-haystack
xapian-haystack
-e git+git://github.com/piquadrat/django-cms-search.git#egg=django-cms-search
-e git://github.com/evdb/django-date-extensions.git#egg=django-date-extensions
django-markitup
Markdown>=2.0.3

# mysociety.org
django-tastypie
# use shipped copy of popit
#-e git+git://github.com/mysociety/popit.git@popit-v0.1dev#egg=popit
#django-mapit



#django-ajax-selects>=1.2.1
